refactored uiRunImage, uiShowReadMe
